requirements.txt
boto==2.39.0
dj-config-url==0.1.1
Django==1.9.4
django-appconf==1.0.1
django-geoposition==0.2.2
django-reversion==1.10.1
Django-Select2==5.8.4
django-storages==1.4
djangorestframework==3.3.3
gunicorn==19.4.5
pbr==1.8.1
Pillow==3.1.1
psycopg2==2.6.1
requests==2.9.1
requests-futures==0.9.7
six==1.10.0
